[ADD] Added module Dev on Sales Dev 1: Add a field on customer record - 'Default Ship Method' This will be a drop-down list of all delivery methods. Add another field ' Customer Account number'. This is the customer delivery a/c detail When creating the SO for the customer, the default ship method should populate under 'Delivery Method' field. Add  ' Customer Account number' to the SO. When creating the SO, this should also be pulled from the contact form if the delivery method is selected.
Dev 2:
Add a new field COO (Country of Origin) on product template.
Add a new field Schedule B Number on product template.
Add a new column COO (Country of Origin) on Sale Order PDF and Delivery Slip PDF and display the relevant COO from the product template ONLY IF the shipping address is outside United States. Add a new field for Schedule B Number and display only if the COO is displayed.
If the country to where the order is being shipped is the United States then this column (COO) and field Schedule B Number should not show up.