[FIX][1885268] MRP Material Shortage Report
The sales people need to check if a product is available or not. The product that they are checking might not have any On Hand qty, but it might have all teh components that are needed to manufacture it.
Hence we need a button called - 'Manufacture Qty' on the product itself. This will display the quantity that can be manufactured (based on the availability of the components and BOM)
Now, the BOM components can also be of type manufacture, so the report should drill down till the last product.

A scheduled action can be created that updates the Manufacture Qty at a certain interval, if the dynamic update will slow down the system.

Sample :

Product - Computer - Current On hand - 0

Components -
Name             Qty         Quantity Available
Screen             1            2
Mouse              1            10
Battery             2            40

Product - Screen - Current Current On hand - 2

Components -
Name             Qty         Quantity Available
LCD Screen          1            100
Keyboard              1            100

So. the manufacture qty for Computer will be 10