[ADD] Blanket Order - The ordered date is right justified. Please make it left justified. (the actual field should be moved to left) PO header should be left justified. The address is messed up. See attached. Remove the word 'Products' Change the footer email to ap@sciencix.com like how it is on the PO.
PO -
Please left justify the Part ID (customer apologizes for the confusion) and the fields.
Decrease the size of the description so that the entire requested date can be in one line.
Remove the dot from Date Req. (It should be Date Req)
Move tax column after unit price ( add it between the unit price and amount)